refactor: sign up enum
