+++ frontend/src/components/auth/SignUp/RegisterForm.tsx
@@ -14,12 +14,12 @@ import Input from '@/components/Primitives/Inputs/Input/Input';
 import Flex from '@/components/Primitives/Layout/Flex/Flex';
 import Text from '@/components/Primitives/Text/Text';
 import { DASHBOARD_ROUTE } from '@/constants/routes';
+import { SignUpEnum } from '@/enums/auth/signUp';
 import { ToastStateEnum } from '@/enums/toasts/toast-types';
 import SchemaRegisterForm from '@/schema/schemaRegisterForm';
 import { toastState } from '@/store/toast/atom/toast.atom';
 import { FlexForm } from '@/styles/pages/pages.styles';
 import { RegisterUser } from '@/types/user/user';
-import { SignUpEnum } from '@/utils/signUp.enum';
 
 interface RegisterFormProps {
   emailName: { email: string; goback: boolean };
