+++ frontend/src/components/auth/SignUp/SignUpForm.tsx
@@ -9,12 +9,12 @@ import Button from '@/components/Primitives/Inputs/Button/Button';
 import Input from '@/components/Primitives/Inputs/Input/Input';
 import Text from '@/components/Primitives/Text/Text';
 import { NEXT_PUBLIC_ENABLE_AZURE } from '@/constants';
+import { SignUpEnum } from '@/enums/auth/signUp';
 import { ToastStateEnum } from '@/enums/toasts/toast-types';
 import SchemaEmail from '@/schema/schemaEmail';
 import { toastState } from '@/store/toast/atom/toast.atom';
 import { FlexForm } from '@/styles/pages/pages.styles';
 import { EmailUser } from '@/types/user/user';
-import { SignUpEnum } from '@/utils/signUp.enum';
 
 interface SignUpFormProps {
   setShowSignUp: Dispatch<React.SetStateAction<SignUpEnum>>;
