+++ frontend/src/enums/auth/signUp.ts
